% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpanm.R
\name{cpanm_install_github}
\alias{cpanm_install_github}
\title{cpanm_install_github}
\usage{
cpanm_install_github(repo, test = FALSE, quiet = TRUE)
}
\arguments{
\item{repo}{The name of the repository - parsed with remotes::parse_github_repo_spec}

\item{test}{Boolean whether to test the dist}

\item{quiet}{Boolean whether to keep quiet}
}
\value{
Boolean
}
\description{
cpanm_install_github
}
