% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perlbrew.R
\name{perlbrew}
\alias{perlbrew}
\title{perlbrew}
\usage{
perlbrew(root = perlbrewr_perlbrew_root(), version = NULL,
  lib = NULL)
}
\arguments{
\item{root}{PERLBREW_ROOT}

\item{version}{perl version}

\item{lib}{local lib name (optional)}
}
\value{
Boolean
}
\description{
perlbrew
}
