% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpanm.R
\name{perlbrew_install_cpanm}
\alias{perlbrew_install_cpanm}
\title{perlbrew_install_cpanm}
\usage{
perlbrew_install_cpanm(force = FALSE)
}
\arguments{
\item{force}{Boolean to force an install/upgrade}
}
\value{
Boolean TRUE for success
}
\description{
perlbrew_install_cpanm
}
