% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perlbrew.R
\name{perlbrew_list}
\alias{perlbrew_list}
\title{perlbrew_list}
\usage{
perlbrew_list(root = Sys.getenv("PERLBREW_ROOT", unset = NA))
}
\arguments{
\item{root}{PERLBREW_ROOT}
}
\value{
character vector
}
\description{
perlbrew_list
}
