% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perlbrew.R
\name{perlbrew_off}
\alias{perlbrew_off}
\title{perlbrew_off}
\usage{
perlbrew_off(root = Sys.getenv("PERLBREW_ROOT", unset = NA))
}
\arguments{
\item{root}{PERLBREW_ROOT}
}
\value{
Boolean
}
\description{
perlbrew_off
}
