% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perlbrew.R
\name{perlbrewr}
\alias{perlbrewr}
\title{perlbrewr}
\usage{
perlbrewr(root = NULL, version = NULL, lib = NULL)
}
\arguments{
\item{root}{PERLBREW_ROOT}

\item{version}{perl version}

\item{lib}{local lib name (optional)}
}
\value{
Boolean
}
\description{
perlbrewr
}
