% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpanm.R
\name{cpanm}
\alias{cpanm}
\alias{cpanm_installdeps}
\title{cpanm}
\usage{
cpanm(installdeps = FALSE, test = TRUE, quiet = TRUE, dist = ".")

cpanm_installdeps(cpanfile = ".")
}
\arguments{
\item{installdeps}{Boolean to use --installdeps}

\item{test}{Boolean to test install}

\item{quiet}{Boolean to run in quiet mode}

\item{dist}{The directory to find a cpanfile, or an arhive file to use}

\item{cpanfile}{Path to cpanfile or a directory to find one}
}
\value{
Boolean TRUE for success
}
\description{
cpanm

cpanm_installdeps
}
