% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpanm.R
\name{cpanm}
\alias{cpanm}
\title{cpanm}
\usage{
cpanm(installdeps = FALSE, test = TRUE, quiet = TRUE,
  cpanfile = ".")
}
\arguments{
\item{installdeps}{Boolean to use --installdeps}

\item{test}{Boolean to test install}

\item{quiet}{Boolean to run in quiet mode}

\item{cpanfile}{THe cpanfile to use}
}
\value{
string
}
\description{
cpanm
}
