% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perlbrew.R
\name{perlbrew_lib_create}
\alias{perlbrew_lib_create}
\title{perlbrew_lib_create}
\usage{
perlbrew_lib_create(root = perlbrewr_perlbrew_root(), version = NULL,
  lib = NULL, perlbrew.use = FALSE)
}
\arguments{
\item{root}{PERLBREW_ROOT}

\item{version}{perl version}

\item{lib}{new local lib name}

\item{perlbrew.use}{switch to the new library.}
}
\value{
Boolean TRUE for success
}
\description{
perlbrew_lib_create
}
