% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perlbrew.R
\name{perlbrew_list}
\alias{perlbrew_list}
\title{perlbrew_list}
\usage{
perlbrew_list(root = perlbrewr_perlbrew_root(), include.libs = TRUE)
}
\arguments{
\item{root}{PERLBREW_ROOT}

\item{include.libs}{whether to include local::lib libraries in output}
}
\value{
character vector
}
\description{
perlbrew_list
}
